Feature/core functions

+++ app/Http/Controllers/UserController.php
@@ -10,24 +10,29 @@ use Auth;
 class UserController extends Controller
 {
     //This function used to get user list
-    public function index()
-    {
+    public function index(Request $request)
+    { 
+        if($request->user()->user_role == 'ADMIN'){
             $users = User::all('name','username','user_role');
             if($users){
-                return response()->json(['status' => 'success','data'=>$users]);
+                return response()->json(['status' => 'success','data'=>$users],200);
             } else {
                 return response()->json(['status' => 'Data is not available'], 200);
             }
+        }else{
+            return response()->json(['errors' => 'Unauthorized Access'], 401);
+        }
     }
 
     //This function used to create user record
     public function create(Request $request)
     {
+        if($request->user()->user_role == 'ADMIN'){
             $validator = Validator::make($request->all(), [
-                'username' => 'required|string|unique:users|min:2|max:10',
+                'username' => 'required|string|unique:users|min:2|max:50',
                 'name' => 'required|string|min:2|max:50',
                 'password' => 'required|min:6|max:10',
-                'user_role'=>'required|string|min:3|max:15'
+                'user_role'=>'required|string|in:ADMIN,PRODUCT_OWNER,TEAM_MEMBER'
             ]);
              
             if ($validator->fails()) {
@@ -40,26 +45,34 @@ class UserController extends Controller
                 'user_role' => $request['user_role'],
             ]);
             return response()->json(['status' => 'User has been created successfully'], 201);
+        }else{
+            return response()->json(['errors' => 'Unauthorized Access'], 401);
+        }
     }
 
     //This function used to get user record by user_id
-    public function getUserById($user_id)
+    public function getUserById(Request $request,$user_id)
     {
+        if($request->user()->user_role == 'ADMIN'){
             $user = User::where('id',$user_id);
             if($user){
-                return response()->json(['status' => 'success','data'=>$user]);
+                return response()->json(['status' => 'success','data'=>$user],200);
             } else {
                 return response()->json(['status' => 'Data is not available'], 200);
             }
+        }else{
+            return response()->json(['errors' => 'Unauthorized Access'], 401);
+        }
     }
 
     //This function used to update user record
     public function update(Request $request, $user_id)
     {
+        if($request->user()->user_role == 'ADMIN'){
             $validator = Validator::make($request->all(), [
-                'username' => 'string|min:2|max:10|unique:users,username,'.$user_id,
+                'username' => 'string|min:2|max:50|unique:users,username,'.$user_id,
                 'name' => 'string|min:2|max:50',
-                'user_role'=>'string|min:3|max:15'
+                'user_role'=>'string|in:ADMIN,PRODUCT_OWNER,TEAM_MEMBER'
             ]);
              
             if ($validator->fails()) {
@@ -69,13 +82,17 @@ class UserController extends Controller
             if($updateUser){
                 return response()->json(['status' => 'User info has been updated successfully'], 200);
             } else {
-                return response()->json(['status' => 'Data is not available'], 200);
+                return response()->json(['status' => 'Data is not updated'], 200);
             }
+        }else{
+            return response()->json(['errors' => 'Unauthorized Access'], 401);
+        }
     }
 
     //This function used to delete user record
-    public function delete($user_id)
+    public function delete(Request $request,$user_id)
     {
+        if($request->user()->user_role == 'ADMIN'){
             $user = User::find($user_id);
             if($user->id){
                 User::where('id',$user_id)->delete();
@@ -83,6 +100,9 @@ class UserController extends Controller
             } else {
                 return response()->json(['status' => 'Data is not available'], 200);
             }
+        }else{
+            return response()->json(['errors' => 'Unauthorized Access'], 401);
+        }
             
     }
 }
